refactor: UI 모델 패키지 위치 변경
- `component` 패키지에서 `model` 패키지로 UI 모델 클래스 이동
- 관련 import 경로 수정
