+++ app/src/androidTest/java/woowacourse/payments/ui/component/CardholderNameTextFieldTest.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.payments.ui.model.CardholderNameUiModel
 
 class CardholderNameTextFieldTest {
     @get:Rule
