+++ app/src/main/java/woowacourse/payments/ui/model/CardPasswordUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.model
 
 import android.os.Parcelable
 import androidx.core.text.isDigitsOnly
