+++ app/src/main/java/woowacourse/payments/ui/screen/registration/CardRegistrationScreenUiState.kt
@@ -3,10 +3,10 @@ package woowacourse.payments.ui.screen.registration
 import android.os.Parcelable
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
-import woowacourse.payments.ui.component.CardExpirationDateUiModel
-import woowacourse.payments.ui.component.CardNumberUiModel
-import woowacourse.payments.ui.component.CardPasswordUiModel
-import woowacourse.payments.ui.component.CardholderNameUiModel
+import woowacourse.payments.ui.model.CardExpirationDateUiModel
+import woowacourse.payments.ui.model.CardNumberUiModel
+import woowacourse.payments.ui.model.CardPasswordUiModel
+import woowacourse.payments.ui.model.CardholderNameUiModel
 
 @Parcelize
 data class CardRegistrationScreenUiState(
