+++ app/src/main/java/woowacourse/payments/ui/screen/registration/CardRegistrationScreenViewModel.kt
@@ -3,11 +3,11 @@ package woowacourse.payments.ui.screen.registration
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import woowacourse.payments.ui.component.CardExpirationDateUiModel
-import woowacourse.payments.ui.component.CardNumberUiModel
-import woowacourse.payments.ui.component.CardPasswordUiModel
-import woowacourse.payments.ui.component.CardholderNameUiModel
-import woowacourse.payments.ui.component.PaymentCardUiModel
+import woowacourse.payments.ui.model.CardExpirationDateUiModel
+import woowacourse.payments.ui.model.CardNumberUiModel
+import woowacourse.payments.ui.model.CardPasswordUiModel
+import woowacourse.payments.ui.model.CardholderNameUiModel
+import woowacourse.payments.ui.model.PaymentCardUiModel
 
 class CardRegistrationScreenViewModel(
     initialUiState: CardRegistrationScreenUiState = CardRegistrationScreenUiState(),
